api/requirements.txt
asgiref==0.10.0
autobahn==0.13.0
channels==0.10.2
daphne==0.10.0
Django==1.9.4
django-cors-headers==1.1.0
djangorestframework==3.3.2
six==1.10.0
Twisted==16.0.0
txaio==2.2.2
zope.interface==4.1.3
